Frontend project creation
Edited front end of the project creation form
